test d'import dynamique pour pas import 27k fichiers et fonctions
